requirements.txt
Django
python-dotenv
GitPython
redis
celery